cache DatabaseServers in `destinations.GetDestination`
GetDestination the caches the DatabaseServer object if successfully opened.

The driver.Conn and sql.DB are stored on the Clickhouse and DubDB
objects respectively since they're not a single connection, but a pool.

MaxIdleConns, and other settings will likely need to be tweaked,
but this prototype should be a fully working demonstration.