implement AWS Redshift destination
it's based on database/sql instead of the AWS SDK mostly because
it was easier to test.
localstack doesn't appear to support Redshift or maybe only in the Pro version.

also, the AWS SDK, AFAICT, isn't designed for persistent connections
so an additional connection pool would be implemented
and database/sql already gives one for free.

configuration for using a local postgres instance:

    [[database_connections]]
    id = "destination-1"
    account_id = "account-1"
    type = "redshift"

      [database_connections.settings]
      tls = false
      endpoint = "http://localhost:5432/scratchdb"
      username = "scratch"
      password = "..."

closes #101