use a slice sqs test cases
previously this test would sometimes fail for unknown reason.

iterating over the map is random so we enqueue messages in random order
and dequeue in random order. if we start dequeue at msg1
and the queue returns msg2, we delete msg2 and the loop stops
because there are no more messages in the map and the test fails.

to simplify everything, just use a slice.